fix: refactor SettingsSampleCodeComposer to be more modular
