+++ src/main/java/com/google/api/generator/gapic/composer/ServiceSettingsClassComposer.java
@@ -49,6 +49,7 @@ import com.google.api.generator.engine.ast.ValueExpr;
 import com.google.api.generator.engine.ast.VaporReference;
 import com.google.api.generator.engine.ast.Variable;
 import com.google.api.generator.engine.ast.VariableExpr;
+import com.google.api.generator.gapic.composer.samplecode.SettingsSampleCodeComposer;
 import com.google.api.generator.gapic.model.GapicClass;
 import com.google.api.generator.gapic.model.GapicClass.Kind;
 import com.google.api.generator.gapic.model.Message;
@@ -130,8 +131,16 @@ public class ServiceSettingsClassComposer implements ClassComposer {
                     .filter(m -> m.stream() == Stream.NONE && !m.hasLro() && !m.isPaged())
                     .findFirst()
                     .orElse(service.methods().get(0)));
+    Optional<String> methodNameOpt =
+        methodOpt.isPresent() ? Optional.of(methodOpt.get().name()) : Optional.empty();
+    Optional<String> sampleCodeOpt =
+        SettingsSampleCodeComposer.composeSampleCode(methodOpt, classType);
     return SettingsCommentComposer.createClassHeaderComments(
-        ClassNames.getServiceClientClassName(service), service.defaultHost(), methodOpt, classType);
+        ClassNames.getServiceClientClassName(service),
+        service.defaultHost(),
+        methodNameOpt,
+        sampleCodeOpt,
+        classType);
   }
 
   private static List<AnnotationNode> createClassAnnotations() {
