+++ src/main/java/com/google/api/generator/gapic/composer/samplecode/BUILD.bazel
@@ -15,6 +15,8 @@ java_library(
     deps = [
         "//src/main/java/com/google/api/generator/engine/ast",
         "//src/main/java/com/google/api/generator/engine/writer",
+        "//src/main/java/com/google/api/generator/gapic/model",
+        "//src/main/java/com/google/api/generator/gapic/utils",
         "@google_java_format_all_deps//jar",
     ],
 )
